Inject reporters into (renamed) AggregateReporter
