+++ core/src/main/scala/stryker4s/report/dashboard/DashboardConfigProvider.scala
@@ -1,7 +1,7 @@
 package stryker4s.report.dashboard
 
 import stryker4s.config.Config
-import stryker4s.env.Environment.Environment
+import stryker4s.env.Environment
 import stryker4s.report.dashboard.Providers._
 import stryker4s.report.model.DashboardConfig
