+++ core/src/main/scala/stryker4s/report/package.scala
@@ -1,0 +1,5 @@
+package stryker4s
+
+package object report {
+  type Reporter = FinishedRunReporter with ProgressReporter
+}
